Add error handling and clean-up some code

+++ space-invaders/hardware.c
@@ -1,13 +1,9 @@
 #include <SDL2/SDL_mixer.h>
 #include "hardware.h"
 
-/* Registers for the IO ports to work with */
-static uint8_t inp1_reg = 0;
-static uint8_t inp2_reg = 0;
-static uint8_t snd1_reg = 0;
-static uint8_t snd2_reg = 0;
-static uint16_t shift_reg = 0;
-static uint8_t shift_amnt_reg = 0;
+#define NUM_SOUNDS 9
+
+static SDL_Event e;
 
 typedef enum {
     CREDIT = 1,
@@ -33,23 +29,34 @@ typedef enum {
     UFO_HIT = (1 << 4)
 } PORT_BITS;
 
+/* Registers for the IO ports to work with */
+static uint8_t inp1_reg = 0;
+static uint8_t inp2_reg = 0;
+static uint8_t snd1_reg = 0;
+static uint8_t snd2_reg = 0;
+static uint16_t shift_reg = 0;
+static uint8_t shift_amnt_reg = 0;
+
+typedef enum {
+    INVADER_DEATH_SND,
+    PLAYER_DEATH_SND,
+    SHOOT_LASER_SND,
+    UFO_HIT_SND,
+    UFO_MOVE_SND,
+    FLEET1_SND,
+    FLEET2_SND,
+    FLEET3_SND,
+    FLEET4_SND
+} SOUND_INST;
+
 /* Sound instances */
-static Mix_Chunk *invader_death_snd;
-static Mix_Chunk *invader_stp1_snd;
-static Mix_Chunk *invader_stp2_snd;
-static Mix_Chunk *invader_stp3_snd;
-static Mix_Chunk *invader_stp4_snd;
-static Mix_Chunk *player_death_snd;
-static Mix_Chunk *shoot_laser_snd;
-static Mix_Chunk *ufo_killed_snd;
-static Mix_Chunk *ufo_move_snd;
+static Mix_Chunk *sounds[NUM_SOUNDS];
 
 // Needed for lame audio bug
 static int ufo_move_chan = -1;
 
 // Sets a pixel of the SDL surface to a certain color.
-static void _set_pixel(SDL_Surface *surface, int x, int y, long color)
-{
+static void _set_pixel(SDL_Surface *surface, int x, int y, long color) {
     Uint32 *pixels = (Uint32 *)surface->pixels;
     pixels[(y * surface->w) + x] = color;
 }
@@ -63,38 +70,38 @@ uint8_t read_inp2(void) {
 }
 
 void write_snd1(uint8_t data) {
-    if (data & UFO) { // This sound keeps playing
-        ufo_move_chan = Mix_PlayChannel(-1, ufo_move_snd, 0);
+    if (data & UFO && sounds[UFO_MOVE_SND]) { // This sound keeps playing
+        ufo_move_chan = Mix_PlayChannel(-1, sounds[UFO_MOVE_SND], 0);
     }
-    if (!(snd1_reg & SHOT) && (data & SHOT)) {
-        Mix_PlayChannel(-1, shoot_laser_snd, 0);
+    if (!(snd1_reg & SHOT) && (data & SHOT) && sounds[SHOOT_LASER_SND]) {
+        Mix_PlayChannel(-1, sounds[SHOOT_LASER_SND], 0);
     }
-    if (!(snd1_reg & PLAYER_DIE) && (data & PLAYER_DIE)) {
-        Mix_PlayChannel(-1, player_death_snd, 0);
+    if (!(snd1_reg & PLAYER_DIE) && (data & PLAYER_DIE) && sounds[PLAYER_DEATH_SND]) {
+        Mix_PlayChannel(-1, sounds[PLAYER_DEATH_SND], 0);
     }
-    if (!(snd1_reg & INVADER_DIE) && (data & INVADER_DIE)) {
-        Mix_PlayChannel(-1, invader_death_snd, 0);
+    if (!(snd1_reg & INVADER_DIE) && (data & INVADER_DIE) && sounds[INVADER_DEATH_SND]) {
+        Mix_PlayChannel(-1, sounds[INVADER_DEATH_SND], 0);
     }
 
     snd1_reg = data;
 }
 
 void write_snd2(uint8_t data) {
-    if (!(snd2_reg & FLEET_MOVE1) && (data & FLEET_MOVE1)) {
-        Mix_PlayChannel(-1, invader_stp1_snd, 0);
+    if (!(snd2_reg & FLEET_MOVE1) && (data & FLEET_MOVE1) && sounds[FLEET1_SND]) {
+        Mix_PlayChannel(-1, sounds[FLEET1_SND], 0);
     }
-    if (!(snd2_reg & FLEET_MOVE2) && (data & FLEET_MOVE2)) {
-        Mix_PlayChannel(-1, invader_stp2_snd, 0);
+    if (!(snd2_reg & FLEET_MOVE2) && (data & FLEET_MOVE2) && sounds[FLEET1_SND]) {
+        Mix_PlayChannel(-1, sounds[FLEET2_SND], 0);
     }
-    if (!(snd2_reg & FLEET_MOVE3) && (data & FLEET_MOVE3)) {
-        Mix_PlayChannel(-1, invader_stp3_snd, 0);
+    if (!(snd2_reg & FLEET_MOVE3) && (data & FLEET_MOVE3) && sounds[FLEET3_SND]) {
+        Mix_PlayChannel(-1, sounds[FLEET3_SND], 0);
     }
-    if (!(snd2_reg & FLEET_MOVE4) && (data & FLEET_MOVE4)) {
-        Mix_PlayChannel(-1, invader_stp4_snd, 0);
+    if (!(snd2_reg & FLEET_MOVE4) && (data & FLEET_MOVE4) && sounds[FLEET4_SND]) {
+        Mix_PlayChannel(-1, sounds[FLEET4_SND], 0);
     }
-    if (!(snd2_reg & UFO_HIT) && (data & UFO_HIT)) {
+    if (!(snd2_reg & UFO_HIT) && (data & UFO_HIT) && sounds[UFO_HIT_SND]) {
         Mix_HaltChannel(ufo_move_chan); // Otherwise kill sound wont play
-        Mix_PlayChannel(-1, ufo_killed_snd, 0);
+        Mix_PlayChannel(-1, sounds[UFO_HIT_SND], 0);
     }
 
     snd2_reg = data;
@@ -110,14 +117,16 @@ void write_shift(uint8_t data) {
 }
 
 void write_shift_amnt(uint8_t data) {
-    shift_amnt_reg = data & 7;
+    shift_amnt_reg = data & 7; // Only the first 3 bits are used
 }
 
 void write_watchdog(uint8_t data) {
-    // No functionality needed
+    /* Space Invaders will call this periodically to reset the watchdog timer.
+     * Since our emulation has no use for a watchdog timer, we can ignore it.
+     */
 }
 
-// This is yucky. Will clean up later.
+// This is yucky. Will clean up later. (Which essentially means I'll never clean it up)
 void display_draw(SDL_Window *window, SDL_Surface *surface, const CPU *cpu) {
     for (int i = 0; i < DISP_BYTES; i++) {
         uint8_t byte = cpu->memory[VIDEO_MEMORY_START + i];
@@ -163,73 +172,99 @@ void display_draw(SDL_Window *window, SDL_Surface *surface, const CPU *cpu) {
     SDL_UpdateWindowSurface(window);
 }
 
-void audio_init(void) {
+bool audio_init(void) {
     if(Mix_OpenAudio(44100, MIX_DEFAULT_FORMAT, 2, 2048) < 0) {
-        printf("SDL_mixer could not initialize! SDL_mixer Error: %s\n", Mix_GetError());
+        return false;
     }
 
-    invader_death_snd = Mix_LoadWAV("../sound/invader_death.wav");
-    invader_stp1_snd = Mix_LoadWAV("../sound/fleet1.wav");
-    invader_stp2_snd = Mix_LoadWAV("../sound/fleet2.wav");
-    invader_stp3_snd = Mix_LoadWAV("../sound/fleet3.wav");
-    invader_stp4_snd = Mix_LoadWAV("../sound/fleet4.wav");
-    player_death_snd = Mix_LoadWAV("../sound/player_death.wav");
-    shoot_laser_snd = Mix_LoadWAV("../sound/shoot_laser.wav");
-    ufo_killed_snd = Mix_LoadWAV("../sound/ufo_hit.wav");
-    ufo_move_snd = Mix_LoadWAV("../sound/ufo_move.wav");
+    sounds[INVADER_DEATH_SND] = Mix_LoadWAV("../sound/invader_death.wav");
+    sounds[FLEET1_SND] = Mix_LoadWAV("../sound/fleet1.wav");
+    sounds[FLEET2_SND] = Mix_LoadWAV("../sound/fleet2.wav");
+    sounds[FLEET3_SND] = Mix_LoadWAV("../sound/fleet3.wav");
+    sounds[FLEET4_SND] = Mix_LoadWAV("../sound/fleet4.wav");
+    sounds[PLAYER_DEATH_SND] = Mix_LoadWAV("../sound/player_death.wav");
+    sounds[SHOOT_LASER_SND] = Mix_LoadWAV("../sound/shoot_laser.wav");
+    sounds[UFO_HIT_SND] = Mix_LoadWAV("../sound/ufo_hit.wav");
+    sounds[UFO_MOVE_SND] = Mix_LoadWAV("../sound/ufo_move.wav");
+
+    for (int i = 0; i < NUM_SOUNDS; i++) {
+        if (!sounds[i]) {
+            fprintf(stderr, "Failed to load sound effect %d\n", i);
+        }
+    }
+
+    return true;
 }
 
 void audio_quit(void) {
     Mix_Quit();
 }
 
-bool handle_input(SDL_Event *e) {
-    while (SDL_PollEvent(e)) {
-        SDL_Keycode keyc = e->key.keysym.sym;
+bool handle_input() {
+    while (SDL_PollEvent(&e)) {
+        SDL_Keycode keyc = e.key.keysym.sym;
 
-        switch (e->type) {
+        switch (e.type) {
         case SDL_QUIT:
             return false;
             break;
         case SDL_KEYDOWN:
-            if (keyc == SDLK_RETURN) {
+            switch (keyc) {
+            case SDLK_RETURN:
                 inp1_reg |= START_1P;
-            } else if (keyc == SDLK_c) {
+                break;
+            case SDLK_c:
                 inp1_reg |= CREDIT;
-            } else if (keyc == SDLK_t) {
+                break;
+            case SDLK_t:
                 inp2_reg |= TILT;
-            } else if (keyc == SDLK_SPACE) {
+                break;
+            case SDLK_SPACE:
                 inp1_reg |= SHOT_1P;
                 inp2_reg |= SHOT_2P;
-            } else if (keyc == SDLK_RIGHT) {
+                break;
+            case SDLK_RIGHT:
                 inp1_reg |= RIGHT_1P;
                 inp2_reg |= RIGHT_2P;
-            } else if (keyc == SDLK_LEFT) {
+                break;
+            case SDLK_LEFT:
                 inp1_reg |= LEFT_1P;
                 inp2_reg |= LEFT_2P;
-            } else if (keyc == SDLK_2) {
+                break;
+            case SDLK_2:
                 inp1_reg |= START_2P;
+                break;
             }
+
             break;
         case SDL_KEYUP:
-            if (keyc == SDLK_RETURN) {
+            switch (keyc) {
+            case SDLK_RETURN:
                 inp1_reg &= ~START_1P;
-            } else if (keyc == SDLK_c) {
+                break;
+            case SDLK_c:
                 inp1_reg &= ~CREDIT;
-            } else if (keyc == SDLK_t) {
+                break;
+            case SDLK_t:
                 inp2_reg &= ~TILT;
-            } else if (keyc == SDLK_SPACE) {
+                break;
+            case SDLK_SPACE:
                 inp1_reg &= ~SHOT_1P;
                 inp2_reg &= ~SHOT_2P;
-            } else if (keyc == SDLK_RIGHT) {
+                break;
+            case SDLK_RIGHT:
                 inp1_reg &= ~RIGHT_1P;
                 inp2_reg &= ~RIGHT_2P;
-            } else if (keyc == SDLK_LEFT) {
+                break;
+            case SDLK_LEFT:
                 inp1_reg &= ~LEFT_1P;
                 inp2_reg &= ~LEFT_2P;
-            } else if (keyc == SDLK_2) {
+                break;
+            case SDLK_2:
                 inp1_reg &= ~START_2P;
+                break;
             }
+
             break;
         }
     }
